+++ examples/chars.rs
@@ -1,13 +1,12 @@
 use std::io::{self, stdout, Read, Write};
 
 use termal::{
-    codes,
     error::Result,
     raw::{disable_raw_mode, enable_raw_mode},
 };
 
 fn main() -> Result<()> {
-    print!("a{}", codes::repeat_char!(5));
+    print!("\x1b[5;10r");
     _ = stdout().flush();
     enable_raw_mode()?;